requirements.txt
accelerate==1.12.0
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.14.3
bertopic==0.17.4
bioc==2.1
blinker==1.9.0
boto3==1.42.38
botocore==1.42.38
cachetools==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
conllu==4.5.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
Deprecated==1.3.1
docopt==0.6.2
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.18.0
fitz==0.0.1.dev2
flair==0.15.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
ftfy==6.3.1
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.2
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.10
importlib_resources==6.5.2
intervaltree==3.2.1
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama
langchain-text-splitters==1.1.0
langdetect==1.0.9
langsmith==0.6.6
llvmlite==0.46.0
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.8.0
mpld3==0.5.12
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.4.9
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.5
pathlib==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.5.1
plotly==6.5.2
pptree==3.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
prov==2.0.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyarrow==21.0.0
pydantic==2.11.5
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
Pygments==2.19.2
PyMuPDF==1.26.0
pynndescent==0.6.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytorch_revgrad==0.2.0
pytz==2025.2
pyvis==0.3.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
rdflib==6.3.2
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.7.1
scipy==1.15.3
segtok==1.5.11
sentence-transformers==5.2.2
sentencepiece==0.2.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.41
sqlitedict==2.1.0
stack-data==0.6.3
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformer-smaller-training-vocab==0.4.2
transformers==4.57.6
triton==3.6.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.11
urllib3==2.4.0
uuid_utils==0.14.0
watchdog==6.0.0
wcwidth==0.2.14
Wikipedia-API==0.9.0
wrapt==2.0.1
yarl==1.22.0
zstandard==0.23.0

